test: move sport show functional tests into his own controller
